ROS-2001: [Test] Full pipeline when there is intermittent internet ROS-2002: [Test] Case when batched data is queued at an untenable rate

+++ cloudwatch_logs_common/include/cloudwatch_logs_common/utils/publisher.h
@@ -115,7 +115,7 @@ public:
       return 0;
     }
     int successes = publish_successes_.load();
-    return successes / attempts * 100.0f;
+    return (float) successes / (float) attempts * 100.0f;
   }
 
   /**
